requirements.txt
addict==2.4.0
attrs==25.1.0
blinker==1.9.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
ConfigArgParse==1.7
contourpy==1.3.0
cycler==0.12.1
dash==2.18.2
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
Faker==35.2.0
fastjsonschema==2.21.1
filelock==3.17.0
Flask==3.0.3
fonttools==4.55.6
freetype-py==2.5.1
fsspec==2024.12.0
glfw==2.8.0
hsluv==5.0.4
idna==3.10
imgui==2.0.0
importlib_metadata==8.6.1
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.5
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_core==5.7.2
kiwisolver==1.4.7
MarkupSafe==3.0.2
matplotlib==3.9.4
mpmath==1.3.0
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.2.1
numpy==1.26.4
open3d==0.18.0
packaging==24.2
pandas==2.2.3
pillow==11.1.0
platformdirs==4.3.6
plotly==5.24.1
PyOpenGL==3.1.9
pyparsing==3.2.1
PyQt5==5.15.11
PyQt5-Qt5==5.15.16
PyQt5_sip==12.16.1
pyquaternion==0.9.9
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
retrying==1.3.4
rpds-py==0.22.3
scikit-learn==1.6.1
scipy==1.13.1
six==1.17.0
sympy==1.13.1
tenacity==9.0.0
threadpoolctl==3.5.0
torch==2.6.0
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
vispy==0.14.3
Werkzeug==3.0.6
zipp==3.21.0
